--- android/app/src/main/java/com/yagubogu/data/datasource/StadiumDataSource.kt
@@ -1,7 +1,0 @@
-package com.yagubogu.data.datasource
-
-import com.yagubogu.data.dto.response.StadiumsResponse
-
-interface StadiumDataSource {
-    suspend fun getStadiums(): Result<StadiumsResponse>
-}
