+++ android/app/src/main/java/com/yagubogu/data/datasource/stadium/StadiumDataSource.kt
@@ -1,0 +1,7 @@
+package com.yagubogu.data.datasource.stadium
+
+import com.yagubogu.data.dto.response.stadium.StadiumsResponse
+
+interface StadiumDataSource {
+    suspend fun getStadiums(): Result<StadiumsResponse>
+}
