+++ android/app/src/main/java/com/yagubogu/data/dto/response/stadium/StadiumsResponse.kt
@@ -1,4 +1,4 @@
-package com.yagubogu.data.dto.response
+package com.yagubogu.data.dto.response.stadium
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
